refactor-track-tbt - 리펙토링 수정본
