+++ src/app/tracks/[id]/page.tsx
@@ -1,5 +1,3 @@
-// import getTrackId from '@/features/tracks/hooks/getTrackId';
-//  import getTrackIdAlbum from '@/features/tracks/hooks/getTrackIdAlbum';
 import { getBaseUrl } from '@/lib/utils/baseUrl';
 import HeaderMain from '@/shared/components/HeaderMain/HeaderMain';
 import TrackDescription from '@/features/tracks/components/TrackDescription/TrackDescription';
